+++ slack/install.sh
@@ -1,7 +1,8 @@
 #!/usr/bin/env bash
 source ./script/lib
 
-set -e
-
-info "Installing slack"
-brew cask install slack
+if test ! -e "/Applications/Slack.app"
+then
+    info "Installing slack"
+    brew cask install slack
+fi
